modify login/logout routes
