+++ lib/middleware/index.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = function(app) {
+
+  app.use(function(req, res, next) {
+
+    app.locals = {
+      user: req.user
+    };
+
+    next();
+  });
+
+};
