+++ lib/routes/schedule/schedule.controller.js
@@ -24,7 +24,7 @@ exports.getAllSchedules = function(req, res) {
 exports.getSchedule = function(req, res) {
   Promise.all([
     new Schedule({id: req.params.id})
-      .fetch({withRelated: ['user', 'courseOfferings', 'extraCurriculars']}),
+      .fetch({withRelated: ['user', 'courseOfferings', 'courseOfferings.meetings', 'courseOfferings.course', 'extraCurriculars']}),
     req.user.fetch()
   ])
   .spread(function(schedule, user) {
