+++ lib/server.js
@@ -50,6 +50,7 @@ app.use(session({
 app.use(passport.initialize());
 app.use(passport.session());
 
+require('./middleware')(app);
 require('./routes')(app);
