+++ lib/views/index.hbs
@@ -1,6 +1,5 @@
 <div id="main-app"></div>
 
-
 {{> course-list-item.tmpl }}
 {{> schedule-list.tmpl }}
 {{> schedule-course-item.tmpl }}
