+++ lib/views/partials/scripts.hbs
@@ -16,11 +16,10 @@
 <!-- views -->
 <script src="./js/views/schedule-main-view.js"></script>
 <script src="./js/views/schedule-view.js"></script>
-<script src="./js/views/schedule-course-view.js"></script>
+<script src="./js/views/schedule-course-item-view.js"></script>
 <script src="./js/views/schedule-list-view.js"></script>
 <script src="./js/views/course-list-item-view.js"></script>
 <script src="./js/views/search-view.js"></script>
-<!-- <script src="./js/views/coursedisplay-view.js"></script> -->
 <script src="./js/views/coursedisplay-view.js"></script>
 <script src="./js/views/schedule-list-item-view.js"></script>
