+++ lib/views/templates/schedule-course-item.tmpl.hbs
@@ -1,6 +1,4 @@
 <script type="text/template" id="schedule-course-item-template">
-  <div class="schedule-course">
-    <h4 class="course-title"></h4>
-
-  </div>
+  <h4 class="list-group-item-heading"><%- course.title %></h4>
+  <p class="list-group-item-text"><%- course.type %></p>
 </script>
